Book of Accounts JavaScript
This project has been created using JavaScript in which we can calculate our daily expenses. By entering the name of the product and its price and on which date it is purchased, we can see its total price and complete table and pdf. Can be downloaded in excel,copy,csv format.